+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.0.4] (2025-03-20)
+
+* Added tests for conditional selection and joint analysis.
+* Fixed bug in joint statistics calculation.
+* Added tests for conditional selection and joint analysis.
+
 ## [0.0.3] (2025-03-20)
 
 * Fixed bug in joint statistics calculation.